--- app/src/pages/inside/dashboardPage/widgets/mostFailedTests/mostFailedTestsTable/count.jsx
@@ -1,20 +1,0 @@
-import * as React from 'react';
-import { number } from 'prop-types';
-import { FormattedMessage } from 'react-intl';
-import classNames from 'classnames/bind';
-import styles from '../mostFailedTests.scss';
-
-const cx = classNames.bind(styles);
-
-const Count = ({ count, total }) => (
-  <div className={cx('count')}>
-    {count} <FormattedMessage id="Common.of" defaultMessage="of" /> {total}
-  </div>
-);
-
-Count.propTypes = {
-  count: number.isRequired,
-  total: number.isRequired,
-};
-
-export default Count;
